+++ AlgorithmComparison/src/main/java/algorithmcomparison/algorithms/MergeSort.java
@@ -2,7 +2,7 @@ package algorithmcomparison.algorithms;
 
 public class MergeSort implements SortAlgorithm {
 
-	private int runtimeMilliseconds;
+	private long runtimeMilliseconds;
 	private int numberOfKeyComparisons;
 	
 	public MergeSort() {
@@ -14,14 +14,14 @@ public class MergeSort implements SortAlgorithm {
 		System.arraycopy(array, 0, tempArray, 0, array.length);
 		final long startTime = System.currentTimeMillis();
 		mergeSort(tempArray);
-		runtimeMilliseconds = (int) (System.currentTimeMillis() - startTime);
+		runtimeMilliseconds = System.currentTimeMillis() - startTime;
 	}
 
 	public int getNumberOfKeyComparisons() {
 		return numberOfKeyComparisons;
 	}
 
-	public int getRuntimeMilliseconds() {
+	public long getRuntimeMilliseconds() {
 		return runtimeMilliseconds;
 	}
